Update PISM to version 2.2.2 and use netcdf-mpi instead of netcdf
This avoid the conflict between the "netcdf" formula and the "petsc"
formula. netcdf depends on "hdf5" and petsc depends on "hdf5-mpi". We
can't install both hdf5 and hdf5-mpi, so it is impossible to install
any program that depends on both netcdf and petsc.

Now PISM depends on netcdf-mpi and petsc, which is OK.